Make tempTouchState a local variable
Currently, mTempTouchState is a class variable, and it is confusing when
it is being modified in functions such as findTouchedWindowAt.

This variable is only used inside findTouchedWindowTargets, so convert
it to a local variable. This will make it easier to keep track of its
contents.

Bug: 142017994
Test: none (this is a cherry-pick)
Change-Id: I8b4f45eadf0ba05e5979de39f78ba81fe79bce31
Merged-In: I8b4f45eadf0ba05e5979de39f78ba81fe79bce31