InputDispatcher: Don't occlude windows with the same PID.
First a review of history. In previous releases there was no PID
consideration in occlusion detection. In a recent patch set we
introduced sending InputChannel-less surfaces from SurfaceFlinger
to InputDispatcher for more complete occlusion detection. As part
of this process we introduced the ownerPid check that is modified
in this CL. We added this because otherwise SurfaceView would begin
occluding their parents, and this would be a compatibility break (and
also not very useful). When adding this line we also added the
token == null check. This was to preserve a previous semantic
where windows from the same ownerPid would occlude eachother in
previous releases. There was no known dependence on this semantic,
it was just easier to add the line to preserve the old behavior than
think about changing it. In light of the linked bug, we see a use case
for changing it. On further consideration, it seems not a useful thing
to preserve, as there are no security boundaries within a process
and the ability for an app to truly depend on this behavior is
very small.

Bug: 157772682
Test: Existing tests pass
Change-Id: Ife2a5a9fce35e04bb794d79b3a1b875859deb8dd